refactor: split SSE task-event follower into gateway/task_events.py
gateway/tasks.py grew to 1726 lines after the P2 slice/discovery work and
tripped the 1600-line module gate (test_smoke::test_no_oversized_modules;
the file is not grandfathered). Move the coherent SSE piece — api_task_events,
_TaskEventFollower, _read_live_jsonl_entries, iter_task_events and their
helpers — verbatim into ouroboros/gateway/task_events.py (547 lines;
tasks.py now 1242).

No behavior change. tasks.py re-exports the moved surface, so the router,
CLI, and tests keep importing from gateway.tasks. Monkeypatch pins that
patch collaborators on gateway.tasks (test_perf_budgets:
_read_live_jsonl_entries; test_task_events_sse: load_effective_task_result,
read_json_dict) stay effective: the moved code resolves those symbols
through the gateway.tasks namespace at call time (task_events._tasks_namespace,
lazy import to avoid the module-load cycle).
